- Successfully bundled the renderer into the library
When I see this binding, the first thing comes in mind
is that we need to ship our own renderer for each
backend, but trying to translate the original C
renderer manually and individually is not an efficient
approach which everyone who have their own renderer
might written differently, and they need to update
their own version of renderer with their own code.

With this commit, users now have an option to either
using the original approach, or they can specify
a specific renderer for their project with an enum,
similar to how zgui works.